Improve .sh File for Output/MPDE/mpde-prn Test
This commit improves the .sh file for the Output/MPDE/mpde-prn test.
It fixes the incorrect exit codes if the Xyce run fails.

In the previous version of this test, the contents of the output
files were not being checked.  I assume that was because xyce_verify
did not handle those three MPDE-specific file formats.  So, I changed
the comparator to file_compare.pl, which showed that the test output
did not actually match the gold standards.

I left those comparator changes in this new version but commented
them out.  This lets the test still pass nightly testing while SON
Bug 969 is worked.  This test does check that the output files are
being made with the correct file names.  So, a "pass" is useful but
incomplete.

SON Bug 969