Temporarily remove valgrind from tags list
The shell script associated with
SEMIC_RESISTOR/semic_resistor_step.cir is incompatible with VALGRIND
testing because it neither uses xyce_verify.pl nor has special code to
work around the way that VALGRIND kludges the test process.  But it
has "valgrind" in its tags list, and so it runs during VALGRIND
testing and then fails.

The failure is meaningless --- it is simply breaking because VALGRIND
swaps "valgrind_check" for "xyce_verify" in the script's command line
argument, and then this shell script tries to swap "xyce_verify" for
"file_compare" --- and because the argument doesn't *have*
"xyce_verify" in it it stays as "valgrind_check" and is run with the
wrong command line arguments for that script.

The shell script should either be modified to use
xyce_verify.pl (which can handle .step netlists just fine, so there's
no need to fall back to "file_compare", and *is* compatible with
valgrind_check), or have additional code added to it to handle
VALGRIND testing, or just not be tested with VALGRIND.

In the meantime, since Pete's not around on Friday and won't be back
till two more nightly test cycles run, I'm removing the "valgrind" tag
from this test's tags list so we don't get any more false valgrind
failures.