Test Failed DC Measures for MIN, MAX and PP
This commit adds subtests for FAILED measures for the MIN, MAX
and PP measures for the DC measure mode.  It also tests the
DEFAULT_VAL keyword for those measure types and the DC measure
mode.  The .ms0 file is correct.  The stdout is "slightly wrong".
The measure values are correctly reported as FAILED in the stdout.
However, the measure window information is incorrect in the stdout.
This test will be reported as "passed" for Xyce 6.9.  The issue with
the incorrect measure window info is now SON Bug 1000.

SON Bug 884 and 1000