Improve Raw File Testing
This commit improves the accuracy of the tests for -r and FORMAT=RAW
in Output/AC and Output/DC.  The contents of the header lines
and the values in the data fields are now all checked by the
associated .sh files.  This dictated updates to some of the "gold
standards".  Per SON Bug 461, a Xyce RAW file should not include
the version tag.  I also changed a few of the title lines,
in the gold standards, based on typos in the associated .cir files.

The caveats for these changes are:
     1) similar work is needed on the tests in Output/TRAN.

     2) the updated testing approach, in the .sh files, works if
        the netlist has only one node.  That is sufficient for
        the tests in Output which are about the correctness of the
        formatting.  There are other tests (e.g., SON Bug 746) that
        test more complicated Raw files.

     3) additional work is needed to update the tests for binary
        raw files with .STEP.

SON Bug 1013