Improve Regression Tests in HB
This commit improves the .sh files for the tests in HB.  The .sh
files for those tests now verify the correctness of the output
against all of the existing gold standards.  They also check for
the existenece, or correct "non-existence", of all of the possible
output files.

The gilbert_cell_hb.cir netlist was changed so it that now produces
output that matches the gold standard for the .HB.TD.prn and
.HB.FD.prn files.  The issue is that Xyce's method for specifying
the .options hb_int line changed several releases ago.  So, I could
have changed the gold standard or changed the test's netlist.  I
went with the latter.

A similar clean-up effort should be done for the tests in HB/RC

SON Bug 985