More Improvements to Raw File Testing
This commit improves the accuracy of more tests for -r in Output/AC
and Output/DC.  The contents of the header lines (which are ASCII)
are now checked by the associated .sh files for these three tests.
This dictated updates to the "gold standards" for these tests.
Per SON Bug 461, a Xyce RAW file should not include the version tag.

See git commit a8c8f6c5, from earlier today, about the caveats
on this updated testing approach, and the remaining work for .TRAN
and also for .STEP.

SON Bug 1013