Remove Valgrind Testing for SRN Bug 1962
After discussion with Tom Russo, the fixes for SRN Bug 1962 do
not require Valgrind testing.  The relevant maps do not contain
pointers, and the relevant functions are just copying data between
STL containers.  The fixes also did not add any new accesses to the
relevant maps.

This test did not have a .sh file.  So, it was sufficient to
just remove its tags file, and use the default tags file for
that subdirectory.

SRN 1962