Test that -r Command Line Opt. with .HB is an Error
This commit adds a test for the "temporary patch" for SON Bug 1019.
In Xyce 6.9, a request for -r for a .HB analysis produced no output
files, with no warning/error message.  That case has been changed to
a parsing error, until -r and/or FORMAT=RAW output is actually
supported for .HB analyses.

SON Bug 1019