Test SON Bug 1033 for Global Parameters
This commit adds a test for global parameters that use VT and TEMP.
It then uses those global parameters in instance and model parameters
for the level 1 resistor. It is also important to test that both
TEMP and VT can be use within one global parameter definition.  It
also tests stepping over TEMP.

SON Bug 1033
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1033