Fixed tests to work w/o c++11, removed exclude tag
The main difference between how this new code behaves with and without
C++11 is that with c++11, it generates random distributions using the
C++ <random> library, and without C++11 it generates them using
N_UTL_RandomNumbers.  So, they go thru a slightly differen path in the
code but should be nearly functionally the same.

These test are all similar to the tests for AGAUSS, in that they perform
a sampling run, and re-do the same sampling run using the seed generated
by the first run.  The seed is obtained via "grep" of stdout from the
first run.

Some of my recent hackage appears to have caused the seed from the first
run to be output 2x instead of 1x.  It gets output directly from my new
UQ code and also from the N_UTL_RandomNumber function.

This screwed up the grep, which
caused the CMD variable, used to run Xyce the second time, to be screwed
up.   I've modified all the grep commands to have the "-m 1" option,
which only finds the first instance.  This should work now.  I have
un-excluded the 3 tests that do not rely on gamma functions.