Initial Commit of Regression Tests for MIXED_SIGNAL
This is the initial commit of regression tests for the Mixed Signal
interface for both Python and VPI.  It is tagged so that it should
not run during normal nightly testing yet, since its tags file has
required:mixedsignal.

These tests have the following known issues that need to be
corrected before the Xyce 6.10 release:

  1) it assumes make install builds that was built with reconfigures
as described in the SON Bug 1051 write-up.

  2) the Python test needs to set the LD_LIBRARY_PATH from within
its .sh script.  This is okay on RHEL6 and RHEL7, but likely not
okay on OSX.

  3) the .c files for the VPI tests have a hard-coded path for
the N_CIR_XyceCInterface.h file.  That needs to be made relative
to where the utils/XyceCInterface directory for the build was.

  4) I'm not checking the contents of the .prn file for the test
VPI/runXyceWithDAC.  I need to verify that .prn output matches what
Rich Schiek generated with the Python interface.

  5) I duplicated the Python example in utils/XyceCInterface/example1.
I still need to duplicate the other two Python examples.

  6) I made a copy of utils/XyceCInterface/xyce_interface.py
in the test subdirectory MIXED_SIGNAL/Python.  That is likely a
"BAD approach", as we only want one copy of that file in the
git repos.

SON Bug 1051
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1051