Added new test, fixed an old one(both correlation)
I just added code to test out the use of LHS sampling with the
correlation matrix.  This is the twoParamsCorrelatedLHS.cir test.

Also I changed both of my "twoParamsCorrleated*" tests so that they are
now DC tests rather than transient.  This is to help insure they are
fast.

the "globalParamNorml_OP_MC.cir" test wasn't originally supposed to be
in this commit, but it doesn't hurt.  It is excluded for now, and I am
still working on it.