Add Additional Tests in MIXED_SIGNAL/Python
This commit adds three new tests to MIXED_SIGNAL/Python.  The test
responseReturnCodes verifies the return codes from the Python method
checkResponseVarName() for the cases where the measure names are
valid or invalid.  It also checks the obtainResponse() method for
the case where the measure name is invalid.

The commandLine test verifies the use of Xyce command line options,
such as -o.  This is needed because xyce_interface.py needs to
translate a Python-style array into a C-style array in its
initialize() method.

The noRun test is currently excluded.  It checks the behavior
of xyce_interface.py when its initialize() method returns 2
("DONE").  That could happen with the -norun Xyce command line
option.  This test is currently excluded because xyce_interface.py
segfaults for this test case.

SON Bugs 1051 and 1055
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1051
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1055