Add required:cxx11 tag, remove exclude
I developed this test a long time ago, but it only works
if the version of Xyce under test was compiled using a
C+11 compliant compiler.

The <random> library that is supported by C++11 (and later)
has support for Gamma distributions.   For normal and uniform
distributions, I created functions that would work with or
without the C++11 <random> library.  It helped that there
were functions already present in the Xyce source code
to support both normal and uniform random number sampling,
in the N_UTL_RandomNumbers class, which Tom implemented to support
expression library functions such as AGAUSS.

Gamma distributions aren't the highest priority for
random sampling, I never bothered to implement a non-C++11
version.  So, this test has been excluded until now.