Test SRN Bug 2039 for .MEASURE and .FOUR Lines
The fix for SRN Bug 2039 means that curly braces can now be used
as a substitute for parentheses within Xyce expressions.  Since
that usage is common in HSPICE expressions, this commit adds a
subtest for it to the existing test MEASURE/ExpressionSyntaxes.

I also updated FOURIER/lead_current_power_four.cir to test
SRN Bug 2039, since the FourierManager also looks for Xyce-style
expressions on .FOUR lines.  It uses similar code to the
MeasureManager.

SRN Bug 2039
https://morannon.sandia.gov/bugzilla/show_bug.cgi?id=2039