Remove .OPTIONS ACLIN Lines from Output/SPARAMS Tests
The use of .OPTIONS ACLIN to specify the SPARCALC value, to produce
SPARAM output, was a temporary measure that will be removed before
the Xyce 6.11 release.  So, this commit removes those .OPTIONS lines
from various tests in Output/SPARAMS.  Instead, one of the tests
verifies that explicitly using .LIN SPARCALC=1 does not cause any
parsing problems.

This commit also simplified the test for the absence of an unwanted
warning message, when no FORMAT=<val> is specified on the .LIN line
and Xyce silently defaults to using Touchstone 2 format.

SON Bug 1105
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1105