Updated to use -hspice-ext random
I just modified the Xyce code to toggle the random number functions in
the expression library from the command line.  I have made it so that
these functions do NOT behave as randome by default.    This feels more
consistent with Hspice, which we are moving towards compatibility with.

In order for the following tests to work, it is necessary to modify the
Xyce invocations to include "-hspice-ext random" as command line
arguments.   If these command line arguments are not used, then
functions such as GAUSS, AGAUSS and RAND will simply return the mean
value.

In Hspice, that is what these functions do if there are no sampling
commands in the netlist.  So, this is consistent with that.

Ultimately, the random-related functions in the expression library need
to be connected directly to the new ".SAMPLING" and ".EMBEDDEDSAMPLING"
analysis types.  But that is more work that I can accomplish before the
code freeze.