Make xyce_verify.pl recognize port sources
xyce_verify.pl reads a netlist and constructs a limited view of the
circuit -- notably, it looks at all the sources in the netlist and
keeps a little record of their + and - nodes, so later, if a source is
swept over using a .DC line, it knows what to expect on the .print
line.

This feature was written when only V and I sources existed, and knew
nothing about P sources (port sources, mostly intended for S parameter
use, but valid in other analyses).

This one-character change means a P source should be treated just like
a V for purposes of output/gold-standard validation against the .print
line, which is all this script needs to know about it.  This change
should be sufficient to allow xyce_verify to be used for new
regression tests of the port device in DC analysis.