Add DC and transient tests for port device
The port device behaves as a voltage source in series with an
impedance when DC, AC and transient analysis are used. This
commit adds both the DC and transient tests for the port devices.
The results match Hspice results well. For the transient test,
the default setting used in Hspice generates inaccurate results.
With the higest accuracy setting for runlvl (runlvl is on by default),
the result is inaccurate. To generate accurate results from Hspice,
the delmax option is used. The results with small delmax match
Xyce results of transient tests