Test Use of TEMP in .GLOBAL_PARAM for "Non-Step" Case
The use of TEMP and VT in .GLOBAL_PARAM statements, that are then
used within device instance parameters, was fixed and tested for
the .STEP case for Xyce 6.10.  This commit renames the existing
test to global_params_step.cir.  It then makes global_params.cir
into a test for the "non-step" case.

SON Bug 1033
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1033