Regenerate gold standards for output test
The two tests Output/HB/hb-files and Output/HB/hb-fd-td-diff-formats
are designed to test the correct behavior of the output manager when
various HB outputs are requested.  The test circuit uses a diode,
and bug 1144 SON identified a spice incompatibility in the diode that
was introduced in 2005 and only fixed today.  As a result of this fix,
these two tests started failing, because their gold standards were
generated before the fix and the comparator in use is too unforgiving.

Since these tests are tests of output format and output manager
behavior, the diode behavior in them is unimportant, so I'm simply
correcting the gold standards.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1144