Add a certification test for bug 1144 SON
Bug 1144 SON pointed out that Xyce's diode model had a nasty
discontinuity at higher temperatures, and that SPICE3F5/ngspice did
NOT.

I tracked this down to a Xyce-specific feature (IRF) that had been
added back in 2005 and documented as defaulting to "strict spice
compatibility" even though it incorrectly applied a
temperature-dependent factor to the reverse current, and started to do
so at a temperature-dependent reverse diode voltage.  This single
block of code accounted for both the temperature-dependent magnitude
of the discontinuity *and* the temperature-dependent voltage at which
the discontinuity appeared.

I corrected the code so it falls back to spice compatibility if IRF is
not specified, as it had originally been stated to do in the initial
commit of the feature.

This test runs a simple test netlist through Xyce in a .STEP over
temperature.  It reverse-biases the diode, so tests only the reverse
current expression.  The gold standard for this test was created by
concatenating ngspice output for 11 different temperatures into a
Xyce-formatted ".prn" file, so this test directly verifies that Xyce
is evaluating the diode device in a SPICE-compatible way as intended.

If for any reason it becomes necessary to regenerate the gold standard
for this test, it should be done by running ngspice to create the gold
standard data, not simply by replacing the gold standard with Xyce
output.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1144