Excluding this test for now.
I am excluding this for a couple of reasons.  one is that I appear to
have an indexing mistake in the FOUR-Sens source code that I am still
tracking down.  That is part of the reason for the failure of this test.
The indexing problem only happens when I have several .FOUR statements
with different fundamental frequencies, and I'm attempting to process
sensitivities with FOUR.    In simpler tests it doesn't appear.

The other reason for excluding for now is that this particular circuit
uses devices (the level=3 MOSFET) that don't have analytic derivatives,
and are thus not that accurate.  So I'm creating a better one with a
more accurate device (BSIM3, subject of another bug for this
release), etc.