New test for bug 1121 SON
Bug 1121 - Make it possible to process sensitivities through .FOUR

My previous attempt at a test for this failed badly.  I think there are
a few reasons for this:

(1) I have an indexing bug which shows up with .sens and multiple .four
statements.  So this test is simpler in this respect.

(2) The bad test (now excluded) uses a device which lacks analytic
derivatives.  Also, that test had lax tolerances for time integration.
This version uses the BSIM3 which has analytic derivatives (since I
added them to the bsim3  for bug 1082 SON)

I have also made the test better by putting .FOUR statements in the
circuit which process finite-differenced (via the netlist, not via the
forcefd option) sensitivities.  This gives me some confidence that the
capability gets a decent answer when forced to be as accurate as
possible.   This test doesn't directly compare the .FOUR results of the
FD signals to the .FOUR results of the analytic sensitivities. But I
have manually inspected them and they look good.  And, both signals are
in the gold standard.