update script to use reltol,not abstol(bug1121son)
As this test is comparing FFT's of sensitivities, the numbers tend to be
very big.  As such, absTol is way too small for it to work well for
purposes of comparison.

This commit modifies the test script so that it uses relTol instead.

This is in support of bug 1121 on the SON.
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1121