add another test for bug 1121 son.
This is very similar to the other test of this capability, except that
it has many more .FOUR statements with a variety of frequencies.  Due to
an indexing mistake (now fixed) this combination caused Xyce to
associated the wrong signal with the wrong variable name.  This test
ensures that this is fixed.

Bug 1121 - Make it possible to process sensitivities through .FOUR

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1121