remove v(3) from .FOUR processing (bug1121 son)
This new test is failing on most platforms but all the failures appear
to be in comparing the phase angle for V(3).  All the other signals are
matching, on all the builds I checked.

I threw V(3) in there simply to make the set of .FOUR commands more
complicated but it was a bad choice because it is non-periodic.

So, with this commit I am replacing it with another signal that is
periodic.