add analytical solution test for bug1166son
Pete added a lot of outputter tests for AC-SENS.  But there also needs
to be tests that assure that it gets the right answer.  That's where
this test (and possibly others) come in.

Per the name, this is a simple low-pass RC filter with a frequency
dependent AC voltage source.   The test has Xyce compute sensitivities
of the output voltage, v(2) w.r.t. the parameters R,C and Vin.

The perl script computes all these same sensitivities analytically.

As the solution is complex, the new sensitivity analysis will
compute sensitivities of VR,VI,VM, and VP, so that means (for 3
parameters) a total of 12 sensitivities.