Modified these to support bug 1088 SON
Bug 1088 - add model class function "setupBaseInstanceContainer" in the ADMS back-end for Xyce
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1088

The point of this bug was to test a new way of doing finite differences,
that required much less storage for transient adjoints.  They should
produce exactly the same answers as the old ones.  To add more tests for
this, I took the existing "FD" netlist for the psp, and changed the
options parameter "forcefd" to "forcedevicefd", which forces the new
version to be used.