Clean-up C-code for MIXED_SIGNAL/VPI/runXyceWithDAC
This commit cleans up the C-code for the regression test
MIXED_SIGNAL/VPI/runXyceWithDAC.  It's mostly variable name
changes to make it a better example.  The test also only worked
because the number of ADCs in the netlist was greater than the
number of DACs.  So, the code is "more correct" now.

SON Bug 1181
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1181