Test That -hspice-ext Supports AND and OR in .FUNC
This commit updates the existing test PARSER/hspiceMath.cir to
verify that the -hspice-ext all or -hspice-ext math command line
options support the use of the HSPICE logical (AND) and logical (OR)
operators (which are && and ||) within .FUNC statements.  It also
now verifies that the HSPICE ^ operator (as an exponentiation synonym)
and the HSPICE logarithm functions also work in .FUNC statements.

SON Bug 1196
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1196