Add tests of binning for BSIM6, PSP102
I added binning dummy parameters to BSIM6, PSP102, EKV2.6, EKV3, and
MVSG-HV.

We have no good existing tests for MVSG-HV, and the model has not yet
been verified, so I'm not testing its binning here.

EKV is not an open model and will not likely be getting used by the
people who need model binning, so while I have implemented it, I'm not
adding tests for it.  That leaves BSIM6 and PSP102.

The three tests here are copies of existing BSIM6 and PSP102 test
cases from their respective directories, modified so that there are
two models for each mosfet, only one of which is correct.  We are
relying on binning to select the correct one.  If it did not, the
tests would fail dramatically.

Because a previous error in binning allowed the first model card to be
picked up even when binning was not correct, these tests make sure
that the correct one is the SECOND model of each batch.

Both PMOS and NMOS are being tested here.  In the BSIM6, they're
tested separately, but for the PSP102 they're being tested in a single
inverter netlist.  Furthermore, the PSP netlist uses mosfets in a
subcircuit, so we are also testing that binning works even when the
MOS instances are not at the same level of circuit hierarchy as the
model cards.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1198