Add correct tag to get this test to run
I forgot to add a "nightly" tag to this tag file, and so it would not
run when the tags list contained "+nightly" (as it does for
continuous and nightly testing).  It would only have run weekly, when
we use the tag "?weekly?nightly" (which lets through tests that are
tagged with neither, too).

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1198