Add a Test for SON Bug 1197
This commit adds a regression test for SON Bug 1197, by using
an expression for the device temperature for both an R and C device.
The gold standard was generated by running the netlist with the
R1:R and C1:C values manually set, in the netlist, to the values
that would apply at 35C in the test netlist.  So, the gold standard
used this:

  R1  1  1a 60
  C1  0  1a C={40u*(1+V(1))}

SON Bug 1197
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1197