Update CAPACITOR/ErrorMessageTest
This commit updates the existing test CAPACITOR/ErrorMessageTest.
It removes a subtest for an error message that is no longer emitted.
It adds subtests for two other error conditions that are checked
in the constructor for the C-device.  They are: 1) both C and Q
being specified as solution dependent; and 2) Q being specified
as solution dependent with an IC also being given.

SON Bug 1197
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1197