Improve Testing for .DATA with .AC, .DC and .STEP
This commit makes improvements/changes to the tests for .DATA with
.AC, .DC and .STEP.  A summary of the changes is:

  1) It verifies that .DC and .STEP lines that use DATA=<name> take
precedence over .DC and .STEP lines that don't.

  2) It adds an error test for the case where there is no valid
.DATA statement in the netlist.  This is only tested for .AC.

  3) It adds a test for the "corner case" of .DATA and its use in
DC sweeps with two tables.  This is NOT how DATA should be used,
and is NOT recommended.  However Xyce should be robust against this
use case since it does work correctly if the two tables use disjoint
variables.  This is only tested with .DC, and not with .STEP.

SON Bug 1188
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1188