Test .NOISE with .DATA; Add Error Tests for .NOISE
This commit adds a test for the use of .DATA with .NOISE.  It also
adds tests for invalid .NOISE lines.  This includes .NOISE lines
that use data=<name> and lines that do not.  These error tests
are similar to the ones done for .AC lines under SON Bug 1146.

SON Bug 1042
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1042