Test Total Noise Int. Calculation for .DATA with .NOISE
This commit adds subtests for the calculation of the integrals of
the total input noise and the total output noise when DATA=<name>
is used on a .NOISE line.  In the noise_data.cir netlist, the
frequencies given in the DATA table are monotonically increasing. So,
the total noise integrals are calculated. The noise_data_noTNI.cir
netlist then purposefully misorders its DATA table.  The values in
the noise_data_noTNI.cir.NOISE.prn file are correct.  However, a
warning message is emitted during netlist parsing, and the
calculation of those two integrals (and their output to stdout)
is suppressed.  The two .sh files verify that the stdout is correct
for both cases.

SON Bug 1042
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1042