Add Regression Test for SON Bug 1203
This commit adds regression tests for SON Bug 1203, which was
about various uses of the TEMP instance parameter as a dependent
parameter.  The new tests cover the cases of:

  1) the device temperature being set both with .STEP TEMP and
.OPTIONS DEVICE TEMP=<val>, as well as being left at the default
value of 27C.

  2) the use of parameters, global parameters and functions in
the expression for the TEMP instance parameter.

  3) the error messages from attempts to use node voltages,
branch currents or lead currents in the expression for the
TEMP instance parameter.

The tests verify that the temperature-corrected resistance
values are correct by measuring I(Rx) for each subtest.

SON Bug 1203
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1203