Test Y- and Z-Parameter Output in Touchstone Files
This commit adds regression tests for the output of Y-parameter and
Z-parameter data in Touchstone1 and Touchstone2 formatted output
files, created by .LIN analyses.  Because of how the outputters
are coded, tests were needed for 1-port and 2-port analyses for
the Touchstone1 format.  Only a test with a 2-port analysis was
needed for Touchstone2 format.  For simplicity of the .sh files,
separate tests were added for Y- and Z-parameters.  For all four
tests, all three allowed data formats (RI, MA and DB) are tested.
The fact that "S" is the default parameter type is tested.

SON Bug 1214
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1214