add new transient test for embedded sampling
This test is a more difficult test than what has previously been in the
test suite, as it it is of a nonlinear circuit.

This is a CMOS inverter, using the BSIM3 model.  The test relies,
currently, upon the hackEnsembleOutput function, which is basically a
hack that I wrote to get around using the IO package outputters.  It was
meant to be temporary.

Pete will be writing IO package outputters for embedded sampling in the
near future, and when he does, then this test will need to be revised.