new test for transient quad. PCE
This is similar to a transient test I created for embedded sampling.
The main difference is that projection (quadrature) PCE is used to
compute the statistical moment outputs, instead of computing them from a
large number of samples.  Relatedly, the size of the problem is much
smaller as it is equal to the number of required quadrature points.