Clean-up .sh File for Test for SON Bug 1210
This commit removes some unused variables from the .sh file for
the existing test Certification_Tests/BUG_1210_SON/bug1210.  Those
variables were used, in the netlist that this test was copied from,
to compare the test and gold measure output.  That comparison is not
done for this test.

SON Bug 1210
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1210