Add representative tests for PSP103t
I added the self-heating version of PSP103 to Xyce as the level 1031
MOSFET model.  This model is just the very same PSP103 Verilog source
code with a single "`define" added to turn on all the ifdefed
self-heating effects.

Only two of the tests in the CMC test suite actually test the self
heating model itself (and these two tests are just the same test run
at different temperatures), and all the others are just duplicates of
the tests that the non-self-heating model uses (though the results of
the code with self-heating are slightly different than without).
There are something like 50 tests in the CMC test suite with multiple
variants each, and they all pass for all
variants (symmetric/asymmetric, nmos/pmos, with and without self
heating, flipped, and with multiplicity).

In order to have *something* in the test suite to make sure this new
model doesn't actually get borken, I'm adding four new tests.  These
are the "two" self-heating tests, combined into a single test with a
temperature step.  There is one test for each use case for the PSP:
symmetric/asymmetric, and nmos/pmos.

The gold standard is provided by a script that takes the CMC reference
output and massages it into Xyce format, rather than by archiving the
massaged data the way all the other tests in this directory do.  This
way, if ever we get a PSP 103.5 model, it will be that much easier to
update the test (just copy the renamed CMC reference data into this
directory).  It took me a while even to find the script I had used to
create the gold standards for the other tests.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1219