Initial Set of Tests for .PRINT ES
This is the initial set of tests for the use of .EMBEDDEDSAMPLING
and .PRINT ES.  The highlights of the new tests in Output/ES are as
follows:

  1) The cases of two output variables for projection_pce=true and
regression_pce=true are tested for .TRAN by es-proj-pce-tran-prn and
es-regr-pce-tran-prn.  These tests also make FORMAT=CSV and FORMAT=
NOINDEX files, but those output files are not verified yet.

  2) The test es-proj-pce-dc-prn covers one output variable for
projection_pce=true for .DC.

  3) The test es-specific-print-params-prn covers the new print
parameters outputSampleStats=false and outputAllSamples=true, which
are their non-default values.

  4) The parsing error caused by using -r and -a with .EMBEDDEDSAMPLING
and .PRINT ES is tested by dashRNotSupported.

  5) The fact that .PRINT ES FORMAT=<RAW|PROBE|TOUCHSTONE|TOUCHSTONE2>
defaults to FORMAT=<STD> is tested by es-formats-default-to-prn.

I also added the test Output/Dasho/es which verifies that the -o
output comes from the .PRINT ES line, and is in FORMAT=STD.  The
other changes to the .sh files in Output/Dasho were because of a
change in a warning message.

SON Bug 1201
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1201