Test S(), Y() and Z() Operators for AC Measures
This commit adds a test for AC mode support for the RF parameter
operators for the EQN, MAX, MIN and PP measures.  It only tests a
sampling of those operators, SR(1,1), YDB(1,1), SI(1,1), ZP(1,1)
and YM(1,1), for a one-port network.  This is adequate since the
complete set of the RF parameter operators is already tested in
Output/SPARAMS.

It is important to test these operators both within and outside of
expression delimiters.  Also, the netlist does not have a .PRINT AC
line.  This tests that the S and Z parameter matrices get made
properly by the AC object, when the .LIN line only requests Y
parameter output.

SON Bug 1221
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1221