Exclude Output/ES/es-no-pce-coeffs from Nightly Testing
This commit excludes the Output/ES/es-no-pce-coeffs test from
nightly testing.  The purpose of this test is to test the .PRINT ES
line for OUTPUT_PCE_COEFFS=true when projection_pce=false and
regression_pce=false on the .OPTIONS EMBEDDEDSAMPLES line.  That
OUTPUT_PCE_COEFFS=true request should be silently ignored.

The issue is that Output tests usually use a gold standard, but
it's hard to develop one, for what is essentially a "random sampling"
test, that works on all supported platforms.  So, the .sh file
for this test needs to be re-thought, before this test is added
back to nightly testing.

SON Bug 1201
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1201