Test Error Checking for S(), Y() and Z() Operators
This commit adds a test for the error checking for the S(), Y()
and Z() operators for the case where one of the indices is greater
than the number of ports in the netlist.  Note that Xyce errors out
cleanly, but that happens part way through making the output files,
rather than during netlist parsing.

SON Bug 1199
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1199