Initial Set of Tests for YLIN Device
This is the initial set of tests for the YLIN device.  It tests
the error messages from various invalid Touchstone 2 input files.
The YLIN/ylin test is then a "dummy test", since the initial version
of that device functions like a simplified R device with an R-value
equal to the first element in the impedance values listed on the
[Reference] (or Option) line in the Touchstone 2 file.

SON Bug 1215
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1215