Test DC -remeasure with STEPNUM Column in prn File
This commit adds a test for -remeasure for DC mode measures
when the remeasured .prn file has a STEPNUM column as its first
column.

The error msg for MEASURE/REMEASURE/remeasure_tran_analysis_mismatch
changed.  So, that test now flags that the remeasured file does not
have TIME column in it.

SON Bug 1209
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1209