Test .OPTIONS OUTPUT PHASE_OUTPUT_RADIANS
This commit tests .OPTIONS OUTPUT PHASE_OUTPUT_RADIANS=false.
That new option allows the Xyce phase output, for the VP(), IP(),
SP(), XP() and ZP() operators and for .PRINT SENS for a .AC
analysis to be toggled between being output in radians vs. degrees.
The default is still for phase output in radians. So, these versions
of the tests produce output in degrees instead.

SON Bug 1208
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1208