Initial Set of Tests for .PRINT PCE
This is the initial set of tests for the use of .PCE and .PRINT PCE.
The highlights of the new tests in Output/PCE are as follows:

  1) The test pce-dc-prn covers the case of two output variables for
.DC  for FORMAT = STD, CSV and NOINDEX.  It also output all of the
samples for the quadrature points.

  2) The parsing error caused by using -r and -a with .PCE and
.PRINT PCE is tested by dashRNotSupported.

  3) The fact that .PRINT PCE FORMAT=<RAW|PROBE|TOUCHSTONE|TOUCHSTONE2>
defaults to FORMAT=<STD> is tested by pce-formats-default-to-prn.

  4) The test pce-no-output verifies that no output occurs from the
.PRINT PCE line if there is no OUTPUTS= argument on the .OPTIONS PCES
line.

I also added the test Output/Dasho/pce which verifies that the -o
output comes from the .PRINT PCE line, and is in FORMAT=STD.

SON Bugs 1229 and 1231
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1229
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1231