Test Illegal Params for .PCE and .OPTIONS PCES
This commit adds a test for illegal values for ALPHA, BETA and
STD_DEVIATIONS for .OPTIONS PCES. For the gamma distribution, alpha
and beta must be strictly positive.  A standard deviation must be
non-negative for any distribution.  It also adds a test for ORDER
being >= 0 on the .PCE line.

SON Bug 1224
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1224