Fix Output/ES/es-no-pce-coeffs Test
This commit changes the Output/ES/es-no-pce-coeffs test so that it
should now pass consistently.  The purpose of the test is to verify
that Xyce does something reasonable if the .PRINT ES line has
OUTPUT_PCE_COEFFS=true when projection_pce=false and regression_pce=false
on the .OPTIONS EMBEDDEDSAMPLES line.  That request should be silently
ignored.

The new test uses a voltage divider circuit where V(2) is essentially
constant.  So, the values in the Output/ES/es-no-pce-coeff.ES.prn
file should now be essentially constant and identical on all
platforms.

SON Bug 1201
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1201